requirements.txt
# Core AI & ML
openai>=1.12.0
sentence-transformers>=2.2.2
nltk>=3.8.1
numpy>=1.24.0
pandas>=1.5.0
matplotlib>=3.6.0
seaborn>=0.12.0
scikit-learn>=1.3.0
transformers>=4.30.0

# PyTorch CPU-only for deployment (lighter and faster)
torch>=2.0.0,<3.0.0
torchvision>=0.15.0,<1.0.0

# Data Processing
beautifulsoup4>=4.12.0
requests>=2.31.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# File Handling
PyPDF2>=3.0.0
python-docx>=0.8.11

# Crawling & Analysis
trafilatura>=1.6.0
urllib3>=2.0.0

# JSON Processing
jsonschema>=4.17.0

# Async HTTP
httpx>=0.24.0

# Airtop Integration
airtop-client>=0.2.2

# Visualization
plotly>=5.15.0
wordcloud>=1.9.2

# Progress Bars
tqdm>=4.65.0

# API Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Async Support
asyncio-mqtt>=0.13.0

# Web Interface
streamlit>=1.28.0
streamlit-option-menu>=0.3.6
streamlit-aggrid>=0.3.4
streamlit-plotly-events>=0.0.6

# Additional UI Components
pillow>=10.0.0
altair>=5.1.0